# Refactored tree receiver