@bmf-san 2022-11-13
Close because it has not been improved.

```sh
go test -bench . -count 1 > new.out
benchstat old.out new.out
name          old time/op  new time/op  delta
Static1-8     31.6ns ± 0%  33.3ns ± 0%   ~     (p=1.000 n=1+1)
Static5-8     80.9ns ± 0%  79.5ns ± 0%   ~     (p=1.000 n=1+1)
Static10-8     128ns ± 0%   138ns ± 0%   ~     (p=1.000 n=1+1)
WildCard1-8    243ns ± 0%   268ns ± 0%   ~     (p=1.000 n=1+1)
WildCard5-8    749ns ± 0%   805ns ± 0%   ~     (p=1.000 n=1+1)
WildCard10-8  1.36µs ± 0%  1.41µs ± 0%   ~     (p=1.000 n=1+1)
Regexp1-8      239ns ± 0%   250ns ± 0%   ~     (p=1.000 n=1+1)
Regexp5-8      768ns ± 0%   742ns ± 0%   ~     (p=1.000 n=1+1)
Regexp10-8    1.38µs ± 0%  1.39µs ± 0%   ~     (p=1.000 n=1+1)
```